ENG-222: add keycloak config info to readme, improve keycloak.json fallback, renamed useKeycloak to keycloakEnabled so it doesn't look like a react hook
